+++ src/ode/mod.rs
@@ -3,6 +3,7 @@ pub(crate) mod explicit_rk;
 pub mod predictions;
 pub(crate) mod rosenbrock;
 pub mod solver;
+pub mod stiffness;
 
 pub use predictions::*;
 pub use solver::*;